chore: export dereference-pipelines methods to utils file
